messaging_app/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
PyJWT==2.9.0

# Optional but recommended for production
sqlparse==0.5.2
tzdata==2024.1
